ci(tox): add lint-fix env and set posargs to smarter defaults
Since linting and testing paths are not specified by default in tox.ini,
they would find various untracked files in my working directory and
cause the tests to fail. This change sets the posargs to a more
reasonable default, and adds a lint-fix environment to tox.ini
which will run black and have it auto format the code.