refactor: improve base/core installation for snaps and add tests
Instead of checking if a core is already on the filesystem when each
snap got installed, do collect all the snaps installed at the end and
iterate over the required bases (eg. core, core22, ...). That makes
the whole process less error prone and also gives the ability to
install a specific core from specific channels or revisions.